requirements.txt
pyyaml
pytest
streamlit